LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.double.worker-expected.txt


PASS createImageData(w, h) double is converted to long
